implemented logger

+++ commands/cmd_clear.py
@@ -1,20 +1,27 @@
 import discord
 import asyncio
+from logger import log
 
 
 async def ex(args, message, client, invoke):
-    try:
-        ammount = int(args[0]) + 1 if len(args) > 0 else 2
-    except:
-        await client.send_message(message.channel, embed=discord.Embed(color=discord.Color.red(), description="Please enter another value than %s" % ammount))
-        return
+    author = message.author.name
+    channel = message.channel.__str__()[20:]
+    if author != channel:
+        try:
+            ammount = int(args[0]) + 1 if len(args) > 0 else 2
+        except:
+            await client.send_message(message.channel, embed=discord.Embed(color=discord.Color.red(), description="Please enter another value than %s" % ammount))
+            return
 
-    messages = []
-    async for m in client.logs_from(message.channel, limit=ammount):
-        messages.append(m)
+        messages = []
+        async for m in client.logs_from(message.channel, limit=ammount):
+            messages.append(m)
 
-    await client.delete_messages(messages)
+        await client.delete_messages(messages)
 
-    return_msg = await client.send_message(message.channel, embed=discord.Embed(color=discord.Color.blue(), description="Cleared %s message(s)." % ammount))
-    await asyncio.sleep(4)
-    await client.delete_message(return_msg)
+        return_msg = await client.send_message(message.channel, embed=discord.Embed(color=discord.Color.blue(), description="Cleared %s message(s)." % ammount))
+        await asyncio.sleep(4)
+        await client.delete_message(return_msg)
+    else:
+        await client.send_message(message.author, embed=discord.Embed(color=discord.Color.red(), description="Can't delete direct messages!"))
+        log("Could not clear message(s)!", "error")
